add: context user & user role logic
